requirements.txt
robotframework==7.1.1
robotframework-browser==19.5.0
robotframework-requests==0.9.7
robotframework-databaselibrary==1.4.4
robotframework-seleniumlibrary==6.4.0
robotframework-pabot==4.0.3
allure-robotframework==2.13.5
